Adding Colonise, Merge Fleet and Split Fleet orders.
Adding Colonise, Merge Fleet and Split Fleet orders
Only Split fleet works correctly ATM

Changes to fleet to allow the new orders and get the number of ships in the
fleet

Changes to Object so that I can get parent object.

Changes to order to allow creating and describing the three new orders.