Split connection into a three class hierarchy.
Split connection into a three class hierarchy
Connection is now much more general and I will be able to use it as a base
class for console and serverSocket and therefore remove the current
multi-threading. PlayerConnection is the base class for player's connections,
implementing all the high level functions, and letting subclasses like
PlayerTcpConnection handle things like the network. It also allows adding
the secure and tunnelling versions later.