Tpserver-cpp is now *single threaded*.
Tpserver-cpp is now *single threaded*
The changes to network allowed me to make the console use the select loop
and therefore remove the blocking io thread.
This is good.
I will add the threads back later when and where we need them.
This also allows me to use readline via callbacks for the console.
The main loop is now in the Network class.