Change the <blah>Log settings to log_<blah> and some FileLogger related changes.
If the file the FileLogger is trying to log to doesn't exist, it throws an exception,
which we catch in Logger::reconfigure(). The filename for the FileLogger to log to
is now controlled by logfile_name, and log_file just turns on/off the file logging
to that name.

Set default filename to log to in Setting::setDefaultValues()
.