Added http tunneling socket (untested).
This adds a http tunneling listen socket to the tpserver-cpp server.
This allows clients to use http proxies that don't support https (CONNECT style)
connections.

It is untested, but should work. Might add a http get message as well.

Added the settings that are possible to sample.conf, and cleaned up includes in
net.cpp a little.
.