Added Advertiser, which uses Publishers, such as Avahi is now.
This is a refactoring of the Avahi class to allow multiple "Publishers" to
publish the contact information for the server. The Advertiser controls the 
publishers, keeping the record of the services being published. Avahi is now
a Publisher, and a new publisher for sending the details to the metaserver will 
be written soon.
.