Big Change. Orders now use OrderParameters, and Objects use OrderManager to get the order types by name.
This is a big change. This break current persistence (and says so). But it is much better architecturely
and easier to maintain. It's probably smaller once I remove the unused files. It also means that a game
developer doesn't need to write a ordertype persistence class for every persistence and order (p*o classes).
Once Objects are moved to a similar parameter system (probably TP04) then the game developer will not have
to interact with the persistence method directly at all.
.