Made PlayerTclConn::readFrame non-blocking safe, and kernel buffer underrun safe.
This is trying to fix what was thought to be a non-blocking socket to disconnect bug, but it
turns out that the sockets are blocking. It was a kernel read buffer underrun causing the
problem. This patch allows non-blocking sockets to be used, and fixes the underrun, for
PlayerTcpConn::readFrame (and PlayerHttpConn, which is a subclass that doesn't override
readFrame). Next to fix is verCheck.
.