Require new version of libtprl, use new features and set prompt.
libtprl 0.1.2 allows the server to set whether signals are caught by readline
or not. The signal catching was interfering with another bit of code when
DNS records could not be looked up, making the console fail.
Added redisplay method to Console, to use libtprl's redrawLineForced().
Set a prompt to help people identify the console ("tpserver-cpp>")
.