Fail adding or probing design order if no valid ships in fleet.
Because of the way tpclient-pywx works (and probing, etc), if the fleettype list is
empty, then it's still valid. If there is something in the fleet type list, then
there had better be more than one ship. Zero ships are allowed, if there is atleast
one with at least one.
.